fixion on db

+++ web/js/storage.js
@@ -39,15 +39,60 @@ function updateBestResults(userId, phaseId) {
 
                 simulations = snapshot.val();
             }).then(() => {
-                window.localStorage.setItem('phase1Data', JSON.stringify(simulations));
 
-                let phaseData = Object.values(JSON.parse(window.localStorage.getItem('phase1Data')));
+                if (simulations != null) {
+                    window.localStorage.setItem('phase1Data', JSON.stringify(simulations));
+                    let phaseData = Object.values(JSON.parse(window.localStorage.getItem('phase1Data')));
+                    
+                    let bestTime = 0;
+                    let bestErrors = 0;
+                    phaseData.forEach(phase => {
+                        let phaseTotalTime = (parseInt(phase.time.m) * 600) + parseInt(phase.time.ms) + (parseInt(phase.time.s) * 10);
+    
+                        if (bestTime == 0) {
+                            bestTime = phaseTotalTime;
+                        } else {
+                            if (phaseTotalTime < bestTime) {
+                                bestTime = phaseTotalTime;
+                            }
+                        }
+    
+                        let phaseTotalErrors = 0;
+                        phase.errorCounter.forEach(error => {
+                            phaseTotalErrors = phaseTotalErrors + error.qty;
+                        });
+    
+                        if (bestErrors == 0) {
+                            bestErrors = phaseTotalErrors;
+                        } else {
+                            if (phaseTotalErrors < bestErrors) {
+                                bestErrors = phaseTotalErrors;
+                            }
+                        }
+                    });
+    
+                    firebase.database().ref('users/' + userId + '/phase1').update({ bestTime: bestTime });
+                    firebase.database().ref('users/' + userId + '/phase1').update({ bestErrors: bestErrors });
+    
+                }
 
+            });
+    } else if (phaseId == "2") {
+        firebase.database().ref('users/' + userId + '/phase2' + '/simulations')
+        .once('value')
+        .then(function (snapshot) {
+            simulations = snapshot.val();
+        }).then(() => {
+            if (simulations != null) {
+                window.localStorage.setItem('phase2Data', JSON.stringify(simulations));
+
+                let phaseData = Object.values(JSON.parse(window.localStorage.getItem('phase2Data')));
+    
                 let bestTime = 0;
                 let bestErrors = 0;
                 phaseData.forEach(phase => {
                     let phaseTotalTime = (parseInt(phase.time.m) * 600) + parseInt(phase.time.ms) + (parseInt(phase.time.s) * 10);
-
+    
                     if (bestTime == 0) {
                         bestTime = phaseTotalTime;
                     } else {
@@ -55,12 +100,12 @@ function updateBestResults(userId, phaseId) {
                             bestTime = phaseTotalTime;
                         }
                     }
-
+    
                     let phaseTotalErrors = 0;
                     phase.errorCounter.forEach(error => {
                         phaseTotalErrors = phaseTotalErrors + error.qty;
                     });
-
+    
                     if (bestErrors == 0) {
                         bestErrors = phaseTotalErrors;
                     } else {
@@ -69,49 +114,11 @@ function updateBestResults(userId, phaseId) {
                         }
                     }
                 });
-
-                firebase.database().ref('users/' + userId + '/phase1').update({ bestTime: bestTime });
-                firebase.database().ref('users/' + userId + '/phase1').update({ bestErrors: bestErrors });
-            });
-    } else if (phaseId == "2") {
-        firebase.database().ref('users/' + userId + '/phase2' + '/simulations')
-        .once('value')
-        .then(function (snapshot) {
-            simulations = snapshot.val();
-        }).then(() => {
-            window.localStorage.setItem('phase2Data', JSON.stringify(simulations));
-
-            let phaseData = Object.values(JSON.parse(window.localStorage.getItem('phase2Data')));
-
-            let bestTime = 0;
-            let bestErrors = 0;
-            phaseData.forEach(phase => {
-                let phaseTotalTime = (parseInt(phase.time.m) * 600) + parseInt(phase.time.ms) + (parseInt(phase.time.s) * 10);
-
-                if (bestTime == 0) {
-                    bestTime = phaseTotalTime;
-                } else {
-                    if (phaseTotalTime < bestTime) {
-                        bestTime = phaseTotalTime;
-                    }
-                }
-
-                let phaseTotalErrors = 0;
-                phase.errorCounter.forEach(error => {
-                    phaseTotalErrors = phaseTotalErrors + error.qty;
-                });
-
-                if (bestErrors == 0) {
-                    bestErrors = phaseTotalErrors;
-                } else {
-                    if (phaseTotalErrors < bestErrors) {
-                        bestErrors = phaseTotalErrors;
-                    }
-                }
-            });
-
-            firebase.database().ref('users/' + userId + '/phase2').update({ bestTime: bestTime });
-            firebase.database().ref('users/' + userId + '/phase2').update({ bestErrors: bestErrors });
+    
+                firebase.database().ref('users/' + userId + '/phase2').update({ bestTime: bestTime });
+                firebase.database().ref('users/' + userId + '/phase2').update({ bestErrors: bestErrors });
+            }
+            
         });
     } else if (phaseId == "3") {
         firebase.database().ref('users/' + userId + '/phase3' + '/simulations')
@@ -119,39 +126,42 @@ function updateBestResults(userId, phaseId) {
         .then(function (snapshot) {
             simulations = snapshot.val();
         }).then(() => {
-            window.localStorage.setItem('phase3Data', JSON.stringify(simulations));
+            if (simulations != null) {
+                window.localStorage.setItem('phase3Data', JSON.stringify(simulations));
 
-            let phaseData = Object.values(JSON.parse(window.localStorage.getItem('phase3Data')));
-
-            let bestTime = 0;
-            let bestErrors = 0;
-            phaseData.forEach(phase => {
-                let phaseTotalTime = (parseInt(phase.time.m) * 600) + parseInt(phase.time.ms) + (parseInt(phase.time.s) * 10);
-
-                if (bestTime == 0) {
-                    bestTime = phaseTotalTime;
-                } else {
-                    if (phaseTotalTime < bestTime) {
+                let phaseData = Object.values(JSON.parse(window.localStorage.getItem('phase3Data')));
+    
+                let bestTime = 0;
+                let bestErrors = 0;
+                phaseData.forEach(phase => {
+                    let phaseTotalTime = (parseInt(phase.time.m) * 600) + parseInt(phase.time.ms) + (parseInt(phase.time.s) * 10);
+    
+                    if (bestTime == 0) {
                         bestTime = phaseTotalTime;
+                    } else {
+                        if (phaseTotalTime < bestTime) {
+                            bestTime = phaseTotalTime;
+                        }
                     }
-                }
-
-                let phaseTotalErrors = 0;
-                phase.errorCounter.forEach(error => {
-                    phaseTotalErrors = phaseTotalErrors + error.qty;
-                });
-
-                if (bestErrors == 0) {
-                    bestErrors = phaseTotalErrors;
-                } else {
-                    if (phaseTotalErrors < bestErrors) {
+    
+                    let phaseTotalErrors = 0;
+                    phase.errorCounter.forEach(error => {
+                        phaseTotalErrors = phaseTotalErrors + error.qty;
+                    });
+    
+                    if (bestErrors == 0) {
                         bestErrors = phaseTotalErrors;
+                    } else {
+                        if (phaseTotalErrors < bestErrors) {
+                            bestErrors = phaseTotalErrors;
+                        }
                     }
-                }
-            });
-
-            firebase.database().ref('users/' + userId + '/phase3').update({ bestTime: bestTime });
-            firebase.database().ref('users/' + userId + '/phase3').update({ bestErrors: bestErrors });
+                });
+    
+                firebase.database().ref('users/' + userId + '/phase3').update({ bestTime: bestTime });
+                firebase.database().ref('users/' + userId + '/phase3').update({ bestErrors: bestErrors });
+            }
+            
         });
     }
 }
